Move mixins into their own folder
"Now there are TWO of them!"

Document what properties are required for the mixins to work.
